move slurm_utils and workflow_management - i liked this name, though, to utils submodule
