+++ lstmcpipe/stages/mc_dl2_to_irfs.py
@@ -7,7 +7,7 @@ import shutil
 import logging
 from pathlib import Path
 from ..slurm_utils import SbatchLstMCStage
-from ..workflow_management import save_log_to_file
+from ..utils import save_log_to_file
 from ..io.data_management import check_and_make_dir_without_verification
