+++ lstmcpipe/stages/mc_dl2_to_sensitivity.py
@@ -3,7 +3,7 @@
 import logging
 from pathlib import Path
 from ..slurm_utils import SbatchLstMCStage
-from ..workflow_management import save_log_to_file
+from ..utils import save_log_to_file
 
 
 log = logging.getLogger(__name__)
