+++ lstmcpipe/stages/mc_merge_dl1.py
@@ -2,7 +2,7 @@
 
 import logging
 from pathlib import Path
-from ..workflow_management import save_log_to_file
+from ..utils import save_log_to_file
 from ..slurm_utils import SbatchLstMCStage
 
 log = logging.getLogger(__name__)
